+++ src/fileFrom/UploadFromUrl.ts
@@ -6,17 +6,24 @@ import {ProgressState} from './UploadFrom'
 import {UploadFrom} from './UploadFrom'
 
 export class UploadFromUrl extends UploadFrom {
+  protected request: Promise<UploadcareFile>
+
   readonly data: UrlData
   readonly settings: Settings
+
+  cancel: (() => void)
 
   constructor(data: UrlData, settings: Settings) {
     super()
 
     this.data = data
     this.settings = settings
+    // TODO: Just reject promise and stop check file on CDN
+    this.cancel = function () {}
+    this.request = this.getFilePromise()
   }
 
-  upload(): Promise<UploadcareFile> {
+  private getFilePromise(): Promise<UploadcareFile> {
     const urlPromise = fromUrl(this.data, this.settings)
 
     this.setProgress(ProgressState.Uploading)