+++ src/types.ts
@@ -11,6 +11,21 @@ export type Settings = {
 
 export type FileData = Blob | File | Buffer
 
+export type OriginalImageInfo = {
+  width: number,
+  height: number,
+  format: string,
+  datetimeOriginal: null | string,
+  geoLocation: null | {
+    latitude: number,
+    longitude: number,
+  },
+  orientation: null | number,
+  dpi: null | Array<number>,
+  colorMode: string,
+  sequence?: boolean,
+}
+
 /* TODO Add sourceInfo */
 export type UploadcareFile = {
   uuid: string,
@@ -22,18 +37,5 @@ export type UploadcareFile = {
   cdnUrlModifiers: null | string,
   originalUrl: null | string,
   originalFilename: null | string,
-  originalImageInfo: null | {
-    width: number,
-    height: number,
-    format: string,
-    datetimeOriginal: null | string,
-    geoLocation: null | {
-      latitude: number,
-      longitude: number,
-    },
-    orientation: null | number,
-    dpi: null | Array<number>,
-    colorMode: string,
-    sequence?: boolean,
-  },
+  originalImageInfo: null | OriginalImageInfo,
 }